Tests: Run all test both with and without transactions
